Fix versioning
This fix allows the SWEET_HOME_VERSION env var to be available both in interactive
ash shell and with zsh through .zshenv. It also fixes a bug where the
published block version would be one version behind the latest due to
the docker-compose pointing to the docker image instead of the local
code.

Change-type: patch